Updated scripts for deployment.

+++ seed.js
@@ -1,3 +1,13 @@
+require('dotenv').config()
+
+let dbUrl
+
+if (process.env.ENVIRONMENT === 'production') {
+  dbUrl = process.env.DB_Prod
+} else {
+  dbUrl = process.env.DB_Dev
+}
+
 const { MongoClient } = require('mongodb')
 const fs = require('fs')
 
@@ -5,24 +15,28 @@ const addChamber = function(list, chamber) {
   return list.map(item => ({ ...item, chamber }))
 }
 
-MongoClient.connect('mongodb://localhost/apella', (err, db) => {
+MongoClient.connect(dbUrl, (err, db) => {
   if (err) {
-    console.error(err)
     process.exit(1)
   }
   const representatives = db.collection('representatives')
-  representatives
-    .deleteMany({})
-    .then(() => JSON.parse(fs.readFileSync('senate.json')))
-    .then(parsed => addChamber(parsed, 'Senate'))
-    .then(input => representatives.insertMany(input))
-    .then(() => JSON.parse(fs.readFileSync('house.json')))
-    .then(parsed => addChamber(parsed, 'House'))
-    .then(input => representatives.insertMany(input))
-    .catch(err => {
-      console.log(err)
-      process.exit(1)
-    })
-    .then(() => console.log('Representatives added!'))
-    .then(() => db.close())
+  if (representatives.findOne({})) {
+    console.log('Good to go!')
+    db.close()
+  } else {
+    representatives
+      .deleteMany({})
+      .then(() => JSON.parse(fs.readFileSync('senate.json')))
+      .then(parsed => addChamber(parsed, 'Senate'))
+      .then(input => representatives.insertMany(input))
+      .then(() => JSON.parse(fs.readFileSync('house.json')))
+      .then(parsed => addChamber(parsed, 'House'))
+      .then(input => representatives.insertMany(input))
+      .catch(err => {
+        console.log(err)
+        process.exit(1)
+      })
+      .then(() => console.log('Representatives added!'))
+      .then(() => db.close())
+  }
 })
